🚚 Move entities to their own place
