+++ Mydemenageur.API/Models/Movers/MoverUpdateModel.cs
@@ -6,7 +6,7 @@ using System.ComponentModel.DataAnnotations;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Mydemenageur.API.Models.Users
+namespace Mydemenageur.API.Models.Movers
 {
     public class MoverUpdateModel
     {
